demonstrate binary upload using JAX-RS mode
- evidence that we can already send binaries
- and that openAPI presence does not restrict the user in using the full
blown functionality of the ivy-rest stack (may ignoring the official
specs).
