+++ connectivity/connectivity-demos-test/src_test/rest/bpm/TestDataMapping.java
@@ -7,7 +7,6 @@ import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 
 import com.axonivy.connectivity.rest.DataMappingData;
-import com.axonivy.connectivity.rest.OpenApiData;
 import com.axonivy.connectivity.rest.Post;
 import com.axonivy.connectivity.rest.odata.OData;
 import com.typicode.jsonplaceholder.Address;
@@ -50,27 +49,6 @@ public class TestDataMapping
 
     OData data = result.data().last();
     assertThat(data.getUsers()).hasSize(8);
-  }
-
-  @Test
-  public void openApi_petListing(BpmClient bpmClient)
-  {
-    ExecutionResult result = bpmClient.start().process("rest/openapi/listPets.ivp").execute();
-    assertThat(result).isNotNull();
-
-    OpenApiData data = result.data().last();
-    assertThat(data.getPets()).isNotEmpty();
-    assertThat(data.getPets().get(0).getName()).isNotEmpty();
-  }
-  
-  @Test
-  public void openApi_petQuery(BpmClient bpmClient)
-  {
-    ExecutionResult result = bpmClient.start().process("rest/openapi/login.ivp").execute();
-    assertThat(result).isNotNull();
-
-    long sessionId = Long.parseLong((String)result.workflow().session().getAttribute("pet.session.id"));
-    assertThat(sessionId).as("got session ID from pet store api login").isGreaterThan(1);
   }
   
   @Test
