+++ connectivity/connectivity-demos/processes/rest/openapi.mod
@@ -19,6 +19,12 @@ oi0 @StartRequest f7 '' #zField
 oi0 @EndTask f8 '' #zField
 oi0 @PushWFArc f10 '' #zField
 oi0 @PushWFArc f9 '' #zField
+oi0 @RestClientCall f12 '' #zField
+oi0 @StartRequest f11 '' #zField
+oi0 @EndTask f13 '' #zField
+oi0 @RestClientCall f14 '' #zField
+oi0 @PushWFArc f15 '' #zField
+oi0 @PushWFArc f16 '' #zField
 >Proto oi0 oi0 openapi #zField
 oi0 f0 outLink listPets.ivp #txt
 oi0 f0 inParamDecl '<> param;' #txt
@@ -118,6 +124,68 @@ oi0 f8 305 241 30 30 0 15 #rect
 oi0 f8 @|EndIcon #fIcon
 oi0 f10 248 256 305 256 #arcP
 oi0 f9 79 256 136 256 #arcP
+oi0 f12 clientId ae69ba01-79b7-4dce-9049-900f8f420907 #txt
+oi0 f12 path /store/order/{orderId} #txt
+oi0 f12 method DELETE #txt
+oi0 f12 552 234 112 44 0 -7 #rect
+oi0 f12 @|RestClientCallIcon #fIcon
+oi0 f11 outLink upload.ivp #txt
+oi0 f11 inParamDecl '<> param;' #txt
+oi0 f11 requestEnabled true #txt
+oi0 f11 triggerEnabled false #txt
+oi0 f11 callSignature upload() #txt
+oi0 f11 startName 8.1.3 #txt
+oi0 f11 startDescription 'Use any undocummented feature' #txt
+oi0 f11 caseData businessCase.attach=true #txt
+oi0 f11 @C|.xml '<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
+<elementInfo>
+    <language>
+        <name>upload.ivp</name>
+    </language>
+</elementInfo>
+' #txt
+oi0 f11 @C|.responsibility Everybody #txt
+oi0 f11 49 337 30 30 -25 17 #rect
+oi0 f11 @|StartRequestIcon #fIcon
+oi0 f13 305 337 30 30 0 15 #rect
+oi0 f13 @|EndIcon #fIcon
+oi0 f14 clientId ae69ba01-79b7-4dce-9049-900f8f420907 #txt
+oi0 f14 path /pet/{petId}/uploadImage #txt
+oi0 f14 queryParams 'additionalMetadata="Ivy is everywhere";
+' #txt
+oi0 f14 templateParams 'petId=1;
+' #txt
+oi0 f14 method JAX_RS #txt
+oi0 f14 clientCode 'import com.axonivy.connectivity.rest.client.file.FileUpload;
+import io.swagger.petstore3.client.Pet;
+import javax.ws.rs.core.MediaType;
+import javax.ws.rs.client.Entity;
+
+java.io.File logo = FileUpload.getIvyLogo();
+Pet pet = client.request()
+	.post(Entity.entity(logo, MediaType.APPLICATION_OCTET_STREAM_TYPE), Pet.class) as Pet;
+
+ivy.log.info("Screenshots: "+pet.photoUrls);
+out.pets.add(pet);' #txt
+oi0 f14 bodyInputType RAW #txt
+oi0 f14 bodyMediaType application/octet-stream #txt
+oi0 f14 resultType io.swagger.petstore3.client.ApiResponse #txt
+oi0 f14 clientErrorCode ivy:error:rest:client #txt
+oi0 f14 statusErrorCode ivy:error:rest:client #txt
+oi0 f14 @C|.xml '<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
+<elementInfo>
+    <language>
+        <name>     Upload File
+(programmatic)</name>
+        <desc>Show that the presence of a strictly defined service descriptor is not a blocker to use advanced JAX-RS Api to interact with the remote service.
+This is usefull during prototyping or to overcome technical boundaries caused by weakly defined services.</desc>
+    </language>
+</elementInfo>
+' #txt
+oi0 f14 136 328 112 48 -44 -15 #rect
+oi0 f14 @|RestClientCallIcon #fIcon
+oi0 f15 79 352 136 352 #arcP
+oi0 f16 248 352 305 352 #arcP
 >Proto oi0 .type com.axonivy.connectivity.rest.OpenApiData #txt
 >Proto oi0 .processKind NORMAL #txt
 >Proto oi0 0 0 32 24 18 0 #rect
@@ -130,3 +198,7 @@ oi0 f6 mainOut f10 tail #connect
 oi0 f10 head f8 mainIn #connect
 oi0 f7 mainOut f9 tail #connect
 oi0 f9 head f6 mainIn #connect
+oi0 f11 mainOut f15 tail #connect
+oi0 f15 head f14 mainIn #connect
+oi0 f14 mainOut f16 tail #connect
+oi0 f16 head f13 mainIn #connect
